requirements.txt
fastapi
uvicorn
requests
pydub
PyPDF2
python-dotenv
google-generativeai
